RequestPart 수정

+++ src/main/java/com/runningmate/runningmate/user/controller/UserController.java
@@ -1,5 +1,7 @@
 package com.runningmate.runningmate.user.controller;
 
+import com.runningmate.runningmate.common.exception.NotFoundUserException;
+import com.runningmate.runningmate.common.utils.SessionUtils;
 import com.runningmate.runningmate.user.dto.Request.UserUpdatePasswordRequestDto;
 import com.runningmate.runningmate.user.aop.LoginCheck;
 import com.runningmate.runningmate.user.dto.Request.UserUpdateRequestDto;
@@ -69,6 +71,7 @@ public class UserController {
     @LoginCheck
     @PatchMapping("/{userId}")
     public ResponseEntity<UserInfoResponseDto> modifyUser(@PathVariable("userId") long userId, @RequestBody UserUpdateRequestDto userUpdateRequestDto) {
+        if(SessionUtils.getLoginSessionUserId() != userId) return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         userService.updateUser(userId, userUpdateRequestDto);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -88,6 +91,7 @@ public class UserController {
     @LoginCheck
     @PatchMapping("/{userId}/image")
     public ResponseEntity modifyUserImage(@PathVariable("userId") long userId, @RequestPart("file") MultipartFile multipartFile) {
+        if(SessionUtils.getLoginSessionUserId() != userId) return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         userService.updateUserImage(userId, multipartFile);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -100,7 +104,8 @@ public class UserController {
      */
     @LoginCheck
     @DeleteMapping("/{userId}")
-    public ResponseEntity deleteUser(@PathVariable("userId") long userId, @RequestPart String password ) {
+    public ResponseEntity deleteUser(@PathVariable("userId") long userId, @RequestBody String password ) {
+        if(SessionUtils.getLoginSessionUserId() != userId) return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         userService.deleteUser(userId, password);
         return new ResponseEntity<>(HttpStatus.OK);
     }
